split up basis base
