+++ src/rydstate/basis/basis_sqdt.py
@@ -1,36 +1,9 @@
 from __future__ import annotations
 
-from abc import ABC
-from typing import get_args
-
 import numpy as np
-from typing_extensions import Self
 
-from rydstate.angular.angular_matrix_element import AngularMomentumQuantumNumbers
-from rydstate.rydberg import RydbergStateMQDT, RydbergStateSQDT, RydbergStateSQDTAlkali, RydbergStateSQDTAlkalineLS
-from rydstate.species.species_object import SpeciesObject
-
-
-class BasisBase(ABC):
-    states: list[RydbergStateSQDT | RydbergStateMQDT]
-
-    def __init__(self, species: str | SpeciesObject) -> None:
-        if isinstance(species, str):
-            species = SpeciesObject.from_name(species)
-        self.species = species
-
-    def __len__(self) -> int:
-        return len(self.states)
-
-    def filter_states(self, qn: str, qn_min: float, qn_max: float) -> Self:
-        if qn in get_args(AngularMomentumQuantumNumbers):
-            self.states = [state for state in self.states if qn_min <= state.angular.calc_exp_qn(qn) <= qn_max]
-        elif qn in ["n", "nu"]:
-            self.states = [state for state in self.states if qn_min <= getattr(state, qn) <= qn_max]
-        else:
-            raise ValueError(f"Unknown quantum number {qn}")
-
-        return self
+from rydstate.basis.basis_base import BasisBase
+from rydstate.rydberg import RydbergStateSQDTAlkali, RydbergStateSQDTAlkalineLS
 
 
 class BasisSQDTAlkali(BasisBase):
